Move some Ryanair classes to Ryanair package
