+++ app/src/main/java/com/crewly/roster/ryanair/RyanairRosterParser.kt
@@ -1,4 +1,4 @@
-package com.crewly.roster
+package com.crewly.roster.ryanair
 
 import com.crewly.account.Account
 import com.crewly.activity.ActivityScope
@@ -6,6 +6,7 @@ import com.crewly.duty.Duty
 import com.crewly.duty.RyanairDutyType
 import com.crewly.duty.Sector
 import com.crewly.models.Company
+import com.crewly.roster.Roster
 import dagger.Lazy
 import io.reactivex.Single
 import org.joda.time.DateTime
